Move specs tests to existing test_specs
